Rename scripts and fix bug with df lengths
